More restructuring to ambry-tools
