+++ ambry-tools/src/main/java/com.github.ambry/store/DumpDataHelper.java
@@ -311,73 +311,6 @@ class DumpDataHelper {
   }
 
   /**
-   * Dumps a single record from the log at a given offset and verifies for corresponding values in index
-   * @param randomAccessFile the {@link RandomAccessFile} referring to log file that needs to be parsed
-   * @param blobId the blobId which that is expected to be matched for the record present at
-   *               <code>offset</code>
-   * @param indexValue the {@link IndexValue} that needs to be compared against
-   * @param coveredRanges a {@link Map} of startOffset to endOffset of ranges covered by records in the log
-   * @throws IOException
-   */
-  /*boolean readFromLogAndVerify(RandomAccessFile randomAccessFile, String blobId, IndexValue indexValue,
-      Map<Long, Long> coveredRanges) throws Exception {
-    long offset = indexValue.getOffset().getOffset();
-    try {
-      LogBlobRecordInfo logBlobRecordInfo = readSingleRecordFromLog(randomAccessFile, offset);
-      if (coveredRanges != null) {
-        coveredRanges.put(offset, offset + logBlobRecordInfo.totalRecordSize);
-      }
-      compareIndexValueToLogEntry(blobId, indexValue, logBlobRecordInfo);
-      if (!logBlobRecordInfo.isDeleted) {
-        logger.trace("{}", logBlobRecordInfo.messageHeader + "\n " + logBlobRecordInfo.blobId.getID() + "\n"
-            + logBlobRecordInfo.blobProperty + "\n" + logBlobRecordInfo.userMetadata + "\n"
-            + logBlobRecordInfo.blobDataOutput);
-      } else {
-        logger.trace("{}", logBlobRecordInfo.messageHeader + "\n " + logBlobRecordInfo.blobId.getID() + "\n"
-            + logBlobRecordInfo.deleteMsg);
-      }
-      return true;
-    } catch (IllegalArgumentException e) {
-      logger.error("Illegal arg exception thrown at  " + randomAccessFile.getChannel().position() + ", "
-          + "while reading blob starting at offset " + offset + " with exception: ", e);
-    } catch (MessageFormatException e) {
-      logger.error("MessageFormat exception thrown at  " + randomAccessFile.getChannel().position()
-          + " while reading blob starting at offset " + offset + " with exception: ", e);
-    } catch (EOFException e) {
-      logger.error("EOFException thrown at " + randomAccessFile.getChannel().position() + " ", e);
-      throw (e);
-    } catch (Exception e) {
-      logger.error("Unknown exception thrown " + e.getMessage() + " ", e);
-    }
-    return false;
-  }*/
-
-  /**
-   * Compares values from index to that in the Log
-   * @param blobId the blobId for which comparison is made
-   * @param indexValue the {@link IndexValue} to be used in comparison
-   * @param logBlobRecordInfo the {@link LogBlobRecordInfo} to be used in comparison
-   */
- /* private void compareIndexValueToLogEntry(String blobId, IndexValue indexValue, LogBlobRecordInfo logBlobRecordInfo) {
-    boolean isDeleted = indexValue.isFlagSet(IndexValue.Flags.Delete_Index);
-    boolean isExpired = isExpired(indexValue.getExpiresAtMs());
-    if (isDeleted != logBlobRecordInfo.isDeleted) {
-      logger.error(
-          "Deleted value mismatch for " + logBlobRecordInfo.blobId + " Index value " + isDeleted + ", Log value "
-              + logBlobRecordInfo.isDeleted);
-    } else if (!logBlobRecordInfo.isDeleted && isExpired != logBlobRecordInfo.isExpired) {
-      logger.error(
-          "Expiration value mismatch for " + logBlobRecordInfo.blobId + " Index value " + isExpired + ", Log value "
-              + logBlobRecordInfo.isExpired + ", index TTL in ms " + indexValue.getExpiresAtMs()
-              + ", log Time to live in secs " + logBlobRecordInfo.timeToLiveInSeconds + ", in ms "
-              + TimeUnit.SECONDS.toMillis(logBlobRecordInfo.timeToLiveInSeconds));
-    } else if (!blobId.equals(logBlobRecordInfo.blobId.getID())) {
-      logger.error("BlobId value mismatch for " + logBlobRecordInfo.blobId + " Index value " + blobId + ", Log value "
-          + logBlobRecordInfo.blobId);
-    }
-  }*/
-
-  /**
    * Updates the {@link Map} of blobIds to {@link LogBlobRecord} with the information about the passed in
    * <code>blobId</code>
    * @param blobIdToLogRecord {@link HashMap} of blobId to {@link LogBlobRecord} that needs to be updated with the
